Listas Propias y Edicion
Implementacion de vista de las listas propias de cada usuario y la edicion de las mismas

+++ app/Http/Controllers/UserController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Auth;
-use User;
 
 class UserController extends Controller
 {
@@ -15,7 +14,7 @@ class UserController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('auth')-> only('profile', 'update_profile');
+        $this->middleware('auth')->only('profile', 'update_profile');
     }
 
     /**
@@ -26,25 +25,53 @@ class UserController extends Controller
     public function profile()
     {
         $user = Auth::user();
-        return view('profile/editprofile', ['user' => $user]);
+        $return = [
+            'username' => $user['username'],
+            'avatar' => $user['avatar'],
+            'description' => $user['description'],
+        ];
+        //dd($return['avatar']);
+        return view('profile/editprofile', ['user' => $return]);
     }
 
-    public function update_profile(Request $request) {
-        $this->validate($request, [
-          'avatar' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
-        ]);
-     
+    public function update_profile(Request $request)
+    {
 
-        $filename = Auth::id().'_'.time().'.'.$request->avatar->getClientOriginalExtension();
-        $request->avatar->move(public_path('/uploads/avatars'), $filename);
-     
+        $this->validate($request, [
+            'username' => ['nullable', 'string', 'max:255', 'alpha_num' ,'unique:users'],
+            'avatar' => ['nullable', 'image', 'mimes:jpg,png,jpeg,gif,svg', 'max:2048'],
+            'descripcion' => ['nullable', 'string', 'max:755'],
+        ]);
+
+        $data = $request->all();
         $user = Auth::user();
-        if(strcmp($user->avatar,"default_avatar.png") !== 0){
-            unlink(public_path('/uploads/avatars/'.$user->avatar));
+
+        $modification = false;
+
+        if ($data['username'] != null) {
+            $user->username = $data['username'];
+            $modification = true;
         }
-        $user->avatar = $filename;
-        $user->save();
-     
+
+        if ($request->has('avatar')) {
+            $filename = Auth::id() . '_' . time() . '.' . $request->avatar->getClientOriginalExtension();
+            $request->avatar->move(public_path('/uploads/avatars'), $filename);
+            if (strcmp($user->avatar, "default_avatar.png") !== 0) {
+                unlink(public_path('/uploads/avatars/' . $user->avatar));
+            }
+            $user->avatar = $filename;
+            $modification = true;
+        }
+
+        if ($data['description'] != null) {
+            $user->description = $data['description'];
+            $modification = true;
+        }
+
+        if($modification){
+            $user->save();
+        }
+
         return redirect()->route('user.profile');
-      }
+    }
 }
